Extract classes to top level packages
