--- sdk-springboot/sdk-spring-core/src/main/java/io/dapr/spring/data/repository/query/SimplePropertyPathAccessor.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2024 The Dapr Authors
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package io.dapr.spring.data.repository.query;
-
-import org.springframework.beans.BeanWrapper;
-import org.springframework.data.mapping.PropertyPath;
-import org.springframework.data.util.DirectFieldAccessFallbackBeanWrapper;
-
-public class SimplePropertyPathAccessor<T> {
-
-  private final Object root;
-
-  public SimplePropertyPathAccessor(Object source) {
-    this.root = source;
-  }
-
-  /**
-   * Gets the value of a property path.
-   *
-   * @param path Property path.
-   * @return Value of the property path.
-   */
-  public Object getValue(PropertyPath path) {
-    Object currentValue = root;
-
-    for (PropertyPath current : path) {
-      currentValue = wrap(currentValue).getPropertyValue(current.getSegment());
-      if (currentValue == null) {
-        break;
-      }
-    }
-
-    return currentValue;
-  }
-
-  BeanWrapper wrap(Object o) {
-    return new DirectFieldAccessFallbackBeanWrapper(o);
-  }
-}
